requirements.txt
babel==2.9.0
python-dateutil==2.6.0
flask-moment==0.11.0
flask-wtf==0.14.3
flask_sqlalchemy==2.4.4

flask==1.1.4
flask_migrate== 2.6.0
werkzeug==1.0.1
phonenumbers==8.12.50